requirements.txt
fastapi
uvicorn
python-dotenv
pyodbc
azure-storage-queue==12.11.0